basic transformer implementation without formatting

+++ src/main/java/ru/saltykov/diploma/editing/FormattedToken.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @AllArgsConstructor
 public class FormattedToken {
     private String token;
-    private Long value;
+    private Integer value;
 }
